LayoutTests/http/tests/security/contentSecurityPolicy/default-src-inline-blocked-expected.txt

CONSOLE MESSAGE: line 1: Refused to execute inline script because of Content-Security-Policy.

CONSOLE MESSAGE: line 1: Refused to execute inline script because of Content-Security-Policy.

This test passes if it doesn't alert fail.
